Issue #83 refactor: 확장 함수 패키징
