+++ app/src/main/java/com/drunkenboys/calendarun/util/extensions/ViewExt.kt
@@ -1,4 +1,4 @@
-package com.drunkenboys.calendarun.util
+package com.drunkenboys.calendarun.util.extensions
 
 import android.view.View
 import android.view.animation.Animation
